+++ src/layouts/BasicLayout/BasicLayout.types.ts
@@ -1,4 +1,7 @@
+import { ColorType } from '@styles';
+
 export interface BasicLayoutProps {
 	style?: any;
 	center?: boolean;
+	bg?: keyof ColorType;
 }
