+++ src/screens/DepositScreen/DepositScreen.tsx
@@ -1,52 +1,25 @@
-import React, { useEffect } from 'react';
-import { approvalState } from '@models/deposit';
-import { globalWalletState } from '@stores/WalletStore';
-import { globalDepositState } from '@stores/DepositStore';
+import React from 'react';
 import { Modal, ScreenLoadingIndicator, ModalReusables } from '@components';
-import { useDeposit, useNavigation } from '@hooks';
 import Deposit from './Deposit/Deposit';
 import OpenAave from './OpenAave/OpenAave';
 import { NotAbleToSaveModal } from '../WalletScreen/Modals';
+import { useDepositScreen } from './DepositScreen.hooks';
 
 const DepositScreen = () => {
-	const navigation = useNavigation();
-	const [approved, setApproved] = React.useState<boolean | undefined>(); // transaction amount is approved?
-	const [notAbleToSaveVisible, setNotAbleToSaveVisible] = React.useState(true);
-	const [addFundsVisible, setAddFundsVisible] = React.useState(false);
 	const {
-		market: { tokens = [] }
-	} = globalDepositState().value;
-	const { address } = globalWalletState().value;
-
-	const { ableToDeposit } = useDeposit();
-
-	useEffect(() => {
-		const loadApproved = async () => {
-			if (tokens[0]) {
-				const { isApproved } = await approvalState(address, tokens[0].address);
-				setApproved(isApproved);
-			}
-		};
-		loadApproved();
-	}, []);
+		notAbleToSaveVisible,
+		notAbleToSaveDismiss,
+		addFundsVisible,
+		dismissAddFunds,
+		onAddFunds,
+		ableToDeposit,
+		approved,
+		setApproved
+	} = useDepositScreen();
 
 	if (ableToDeposit === undefined || approved === undefined) {
 		return <ScreenLoadingIndicator />;
 	}
-
-	const notAbleToSaveDismiss = () => {
-		setNotAbleToSaveVisible(false);
-		navigation.goBack();
-	};
-
-	const onAddFunds = () => {
-		setAddFundsVisible(true);
-	};
-
-	const dismissAddFunds = () => {
-		setAddFundsVisible(false);
-		navigation.goBack();
-	};
 
 	if (!ableToDeposit) {
 		return (
